+++ python/qx_utilities/general/utilities.py
@@ -2400,7 +2400,7 @@ def run_recipe(recipe_file=None, recipe=None, steps=None, logfolder=None, eargs=
                 parameters[parameter] = value
     else:
         # define recipe name
-        recipe = "steps parameter"
+        recipe = "steps"
 
         # create the commands dict
         recipe_dict = {}
@@ -2472,10 +2472,6 @@ def run_recipe(recipe_file=None, recipe=None, steps=None, logfolder=None, eargs=
     print(f"---> Running commands from recipe: {recipe}")
     print(f"---> Running commands from recipe: {recipe}\n", file=log)
 
-    if steps:
-        print(f"---> Steps: {steps}")
-        print(f"---> Steps: {steps}\n", file=log)
-
     # commands
     if "commands" not in recipe_dict:
         raise ge.CommandFailed(
@@ -2500,6 +2496,20 @@ def run_recipe(recipe_file=None, recipe=None, steps=None, logfolder=None, eargs=
                 commands_subset.append(com)
 
         commands = commands_subset
+
+    # print commands
+    print("---> Commands:")
+    print("---> Commands:", file=log)
+    commands_set = []
+    for com in commands:
+        if isinstance(com, dict):
+            command_name = list(com.keys())[0]
+        else:
+            command_name = com
+        if command_name not in commands_set:
+            commands_set.append(command_name)
+            print(f"    - {command_name}")
+            print(f"    - {command_name}", file=log)
 
     # XNAT initial setup
     # If running on XNAT, try and load checkpoint if supplied